separate file for Bytes/ByteView
