+++ silkworm/core/common/base.hpp
@@ -21,9 +21,6 @@
 #include <concepts>
 #include <cstddef>
 #include <cstdint>
-#include <span>
-#include <string>
-#include <string_view>
 #include <tuple>
 
 #include <evmc/evmc.hpp>
@@ -42,38 +39,6 @@ using namespace evmc::literals;
 template <class T>
 concept UnsignedIntegral = std::unsigned_integral<T> || std::same_as<T, intx::uint128> ||
     std::same_as<T, intx::uint256> || std::same_as<T, intx::uint512>;
-
-using Bytes = std::basic_string<uint8_t>;
-
-class ByteView : public std::basic_string_view<uint8_t> {
-  public:
-    constexpr ByteView() noexcept = default;
-
-    constexpr ByteView(const std::basic_string_view<uint8_t>& other) noexcept
-        : std::basic_string_view<uint8_t>{other.data(), other.length()} {}
-
-    ByteView(const Bytes& str) noexcept : std::basic_string_view<uint8_t>{str.data(), str.length()} {}
-
-    constexpr ByteView(const uint8_t* data, size_type length) noexcept
-        : std::basic_string_view<uint8_t>{data, length} {}
-
-    template <std::size_t N>
-    constexpr ByteView(const uint8_t (&array)[N]) noexcept : std::basic_string_view<uint8_t>{array, N} {}
-
-    template <std::size_t N>
-    constexpr ByteView(const std::array<uint8_t, N>& array) noexcept
-        : std::basic_string_view<uint8_t>{array.data(), N} {}
-
-    constexpr ByteView(const evmc::address& address) noexcept : ByteView{address.bytes} {}
-
-    constexpr ByteView(const evmc::bytes32& hash) noexcept : ByteView{hash.bytes} {}
-
-    template <std::size_t Extent>
-    constexpr ByteView(std::span<const uint8_t, Extent> span) noexcept
-        : std::basic_string_view<uint8_t>{span.data(), span.size()} {}
-
-    [[nodiscard]] bool is_null() const noexcept { return data() == nullptr; }
-};
 
 using BlockNum = uint64_t;
 using BlockNumRange = std::pair<BlockNum, BlockNum>;
