+++ silkworm/core/common/bytes.hpp
@@ -1,0 +1,59 @@
+/*
+   Copyright 2023 The Silkworm Authors
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+#pragma once
+
+#include <span>
+#include <string>
+#include <string_view>
+
+#include <evmc/evmc.hpp>
+
+namespace silkworm {
+
+using Bytes = std::basic_string<uint8_t>;
+
+class ByteView : public std::basic_string_view<uint8_t> {
+  public:
+    constexpr ByteView() noexcept = default;
+
+    constexpr ByteView(const std::basic_string_view<uint8_t>& other) noexcept
+        : std::basic_string_view<uint8_t>{other.data(), other.length()} {}
+
+    ByteView(const Bytes& str) noexcept : std::basic_string_view<uint8_t>{str.data(), str.length()} {}
+
+    constexpr ByteView(const uint8_t* data, size_type length) noexcept
+        : std::basic_string_view<uint8_t>{data, length} {}
+
+    template <std::size_t N>
+    constexpr ByteView(const uint8_t (&array)[N]) noexcept : std::basic_string_view<uint8_t>{array, N} {}
+
+    template <std::size_t N>
+    constexpr ByteView(const std::array<uint8_t, N>& array) noexcept
+        : std::basic_string_view<uint8_t>{array.data(), N} {}
+
+    constexpr ByteView(const evmc::address& address) noexcept : ByteView{address.bytes} {}
+
+    constexpr ByteView(const evmc::bytes32& hash) noexcept : ByteView{hash.bytes} {}
+
+    template <std::size_t Extent>
+    constexpr ByteView(std::span<const uint8_t, Extent> span) noexcept
+        : std::basic_string_view<uint8_t>{span.data(), span.size()} {}
+
+    [[nodiscard]] bool is_null() const noexcept { return data() == nullptr; }
+};
+
+}  // namespace silkworm
